Refactor Main form into classes per tab
